+++ CHANGELOG.md
@@ -3,6 +3,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [0.1.13] - 2017-11-03
+### Added
+- Added line magic %notify to trigger notifications mid-cell (from @bensteers)
+- Added %autonotify line magic to automatically notify after a certain length of time threshold (from @bensteers)
+- Added --output/-o option to use cell output as message (from @bensteers)
+- Added a jupyter notebook to run through the tests specified in the README
+
+## [0.1.12] - 2017-09-25
+### Added
+- Added some badges to the README and PyPI
+
 ## [0.1.11] - 2017-08-25
 ### Fixed
 - Fixed a small formatting typo in the README